+++ src/test/compile-fail/trait-coercion-generic-bad.rs
@@ -24,7 +24,7 @@ impl Trait<&'static str> for Struct {
 }
 
 fn main() {
-    let s: Box<Trait<int>> = box Struct { person: "Fred" };
-    //~^ ERROR the trait `Trait<int>` is not implemented for the type `Struct`
+    let s: Box<Trait<isize>> = box Struct { person: "Fred" };
+    //~^ ERROR the trait `Trait<isize>` is not implemented for the type `Struct`
     s.f(1);
 }
