+++ src/test/compile-fail/tuple-arity-mismatch.rs
@@ -10,12 +10,12 @@
 
 // Issue #6155
 
-fn first((value, _): (int, f64)) -> int { value }
+fn first((value, _): (isize, f64)) -> isize { value }
 
 fn main() {
     let y = first ((1,2.0,3));
     //~^ ERROR expected a tuple with 2 elements, found one with 3 elements
 
     let y = first ((1,));
-    //~^ ERROR expected `(int, f64)`, found `(int,)`
+    //~^ ERROR expected `(isize, f64)`, found `(isize,)`
 }
